requirements.txt
pygame
numpy
pillow

pyobjc-framework-Cocoa